Move OptOutManager to its own module
